requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
drf-nested-routers==0.93.3
django-cors-headers==3.7.0
channels==3.0.3
channels-redis==3.2.0
django-redis==4.12.1